Refactor search logic into a class
What we had until now was ugly and a tangled mess.

This is a slightly less ugly tangled mess, which is kept out of the way.

It would be nice to replace the SearchBuilder class and instead have the Search
class (currently just a data-carrier) instead have methods which do the work
more or less on-demand.